Use `arc_swap` to implement `AtomicStr` (#72)
* Use `arc_swap` to implement `AtomicStr`

The previous implementation allowed use-after-free in a multi-threaded
context. This PR fixes the problem by using `ArcSwap`, which is
implements thread-safe swapping of `Arc`s, and is widely used.

* Use `triomphe::Arc` to mitigate performance losses

This change replaces `std::sync::Arc` with `triomphe::Arc`. The latter
has no weak references, and is a lot faster because of that.

* Update atomic_str.rs

* Update lib.rs

* Update lib.rs

---------

Co-authored-by: Jason Lee <huacnlee@gmail.com>